hard reload on login logout
