+++ blocks/personalization-card/personalization-card.js
@@ -1,11 +1,4 @@
-function isValidJSON(str) {
-  try {
-    JSON.parse(str);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
+import { isValidJSON } from '../../scripts/personalisation-helpers.js';
 
 function createCardOffer(offer) {
   const offerElement = document.createElement('div');
